% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateManifest.R
\name{CreateManifest}
\alias{CreateManifest}
\title{Create a Manifest of Sound Files}
\usage{
CreateManifest(
  input_dir,
  output_dir,
  year = as.integer(format(Sys.Date(), "\%Y")),
  area_pre = "",
  technician = "",
  recorderType = "SwiftOne",
  output_prefix = "soundManifest_"
)
}
\arguments{
\item{input_dir}{Character. Path to the directory containing `.flac` files.}

\item{output_dir}{Character. Path where the manifest CSV will be saved.}

\item{year}{Integer. Year of recording (default = current year).}

\item{area_pre}{Character. Area/site prefix (default = `""`).}

\item{technician}{Character. Name/ID of technician.}

\item{recorderType}{Character. Recorder type (default = `"SwiftOne"`).}

\item{output_prefix}{Character. Prefix for the output file (default = `"soundManifest_"`).}
}
\value{
A data frame with one row per sound file and the following columns:
\itemize{
  \item \code{path} – relative directory of file
  \item \code{file} – file name
  \item \code{subsequent.file1}, \code{subsequent.file2} – possible subsequent files
  \item \code{FileSize.MB}, \code{FileType}, \code{recorderType}, \code{area}, \code{group}, \code{plot}
  \item \code{plot.alias}, \code{recorder}, \code{recorderAlias}, \code{extraAttrib1}, \code{extraAttrib2}
  \item \code{year}, \code{technician}, \code{date.mmdd}, \code{startTime.hhmm},
  \item \code{startTime.Excel}, \code{startTime.yyyymmddhhmm}, \code{fileLength.min}
  \item \code{MD5_checksum}
}
}
\description{
This function scans a directory of .flac sound files and creates a manifest with metadata including file size, duration, time stamps (parsed from file
names), plot/group identifiers, and MD5 checksums. It assumes a file naming convention where: Date-time stamp starts 20 characters from the end of the filename. Group ID = first 4 characters of the filename. Plot ID = characters 5–8 of the filename.Files are in chronological order.
}
\examples{
\dontrun{
CreateManifest(
  input_dir = "G:/HB",
  output_dir = "C:/output",
  year = 2023,
  area_pre = "NPS",
  technician = "PoojaP"
)
head(manifest)
}

}
